增加接口

+++ src/main/java/com/spider/controller/VideoController.java
@@ -50,7 +50,7 @@ public class VideoController extends BaseController {
     public ResponseVo<Object> cleanVideo(@RequestParam(required = false, defaultValue = "640") Integer height, @RequestParam(required = false, defaultValue = "640") Integer width) {
         AtomicLong size = new AtomicLong(0);
         CopyOnWriteArrayList<String> copyOnWriteArrayList = new CopyOnWriteArrayList<>();
-        videoService.findAll().stream().filter(video -> new File(video.getSavePath()).exists()).filter(video -> Objects.nonNull(video.getMultimediaInfo())).forEach(video -> {
+        videoService.findAll().stream().filter(video -> !video.getSavePath().contains("里番")).filter(video -> new File(video.getSavePath()).exists()).filter(video -> Objects.nonNull(video.getMultimediaInfo())).forEach(video -> {
             if (height * width > video.getMultimediaInfo().getVideo().getSize().getHeight() * video.getMultimediaInfo().getVideo().getSize().getWidth()) {
                 logger.info(video.getSavePath());
                 new File(video.getSavePath()).delete();
@@ -70,8 +70,8 @@ public class VideoController extends BaseController {
     public ResponseVo<Object> cleanTimeVideo(@RequestParam(required = false, defaultValue = "10") Integer minute) {
         AtomicLong size = new AtomicLong(0);
         CopyOnWriteArrayList<String> copyOnWriteArrayList = new CopyOnWriteArrayList<>();
-        videoService.findAll().stream().filter(video -> new File(video.getSavePath()).exists()).filter(video -> Objects.nonNull(video.getMultimediaInfo())).forEach(video -> {
-            if (video.getMultimediaInfo().getDuration()<1000*60*minute) {
+        videoService.findAll().stream().filter(video -> !video.getSavePath().contains("里番")).filter(video -> new File(video.getSavePath()).exists()).filter(video -> Objects.nonNull(video.getMultimediaInfo())).forEach(video -> {
+            if (video.getMultimediaInfo().getDuration() < 1000 * 60 * minute) {
                 logger.info(video.getSavePath());
                 new File(video.getSavePath()).delete();
                 size.addAndGet(video.getSize());
@@ -173,8 +173,10 @@ public class VideoController extends BaseController {
                     videoService.insert(video);
                     logger.info("{},保存完成", video.getSavePath());
                 } else {
-                    logger.info("{}存在的,目标:{}", findVideo.getSavePath(), file.getAbsolutePath());
-                    file.delete();
+                    if (!findVideo.getSavePath().equals(file.getAbsolutePath())) {
+                        logger.info("{}存在的,目标:{}", findVideo.getSavePath(), file.getAbsolutePath());
+                        file.delete();
+                    }
                 }
             }
         });
